DC 입금시 거래 체결 1차 완료

+++ bnk_final_ap/src/main/java/com/example/memo/company/dto/DcAccountDetailsDto.java
@@ -1,5 +1,6 @@
 package com.example.memo.company.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import lombok.Builder;
@@ -10,7 +11,7 @@ public class DcAccountDetailsDto {
     // 계좌 정보
     private String accountNo;
     private String status;
-    private Long balance;
+    private BigDecimal balance;
     private LocalDateTime createdAt;
     
     // 가입자 정보
